+++ sparse_autoencoder/autoencoder/components/tests/test_abstract_encoder.py
@@ -2,7 +2,7 @@
 
 from typing import final
 
-from jaxtyping import Float, Int
+from jaxtyping import Float, Int64
 import pytest
 import torch
 from torch import Tensor
@@ -74,11 +74,11 @@ def test_update_dictionary_vectors_with_no_neurons(mock_encoder: MockEncoder) ->
     torch.random.manual_seed(0)
     original_weight = mock_encoder.weight.clone()  # Save original weight for comparison
 
-    dictionary_vector_indices: Int[Tensor, Axis.INPUT_OUTPUT_FEATURE] = torch.empty(
+    dictionary_vector_indices: Int64[Tensor, Axis.INPUT_OUTPUT_FEATURE] = torch.empty(
         0,
-        dtype=torch.int,  # Empty tensor with 1 dimension
+        dtype=torch.int64,  # Empty tensor with 1 dimension
     )
-    updates: Int[Tensor, Axis.INPUT_OUTPUT_FEATURE] = torch.empty(
+    updates: Float[Tensor, Axis.INPUT_OUTPUT_FEATURE] = torch.empty(
         (0, 0),
         dtype=torch.float,  # Empty tensor with 2 dimensions
     )
@@ -103,8 +103,8 @@ def test_update_dictionary_vectors_with_no_neurons(mock_encoder: MockEncoder) ->
 )
 def test_update_dictionary_vectors_with_neurons(
     mock_encoder: MockEncoder,
-    dictionary_vector_indices: Int[Tensor, Axis.INPUT_OUTPUT_FEATURE],
-    updates: Int[Tensor, Axis.INPUT_OUTPUT_FEATURE],
+    dictionary_vector_indices: Int64[Tensor, Axis.INPUT_OUTPUT_FEATURE],
+    updates: Float[Tensor, Axis.INPUT_OUTPUT_FEATURE],
 ) -> None:
     """Test update_dictionary_vectors with 1 or 2 neurons to update."""
     mock_encoder.update_dictionary_vectors(dictionary_vector_indices, updates)
